requirements.txt
fastapi
uvicorn
python-dotenv
openai
streamlit
requests
pydantic
google-genai